Time: 160 ms (38.49%), Space: 35.5 MB (88.49%) - LeetHub

+++ 1383-maximum-performance-of-a-team/1383-maximum-performance-of-a-team.cpp
@@ -1,13 +1,13 @@
 class Solution {
 public:
     long long mod=1000000007;
-    long long  maxPerformance(int n, vector<int> &speed, vector<int> &efficiency, int k) {
-        vector <pair<long long , long long >> engineers(n);
+    int  maxPerformance(int n, vector<int> &speed, vector<int> &efficiency, int k) {
+        vector <pair<int , int >> engineers(n);
         for (int i = 0; i < n; i++)
             engineers[i] = {efficiency[i], speed[i]};
         sort(engineers.begin(), engineers.end());
         long long  sum = 0, ans = 0;
-        priority_queue<long long > pq;
+        priority_queue<int > pq;
         for (int i = n - 1; i >= 0; i--) {
             sum += engineers[i].second;
             while (pq.size() > k-1) {
